Move backgrounds-related code inside features/backgrounds directory
