+++ src/features/backgrounds/constants.js
@@ -1,0 +1,6 @@
+export const BACKGROUND_AUTOMATIC = 'AUTOMATIC';
+export const BACKGROUND_CUSTOM = 'CUSTOM';
+
+export const BACKGROUND_REFRESH_DAILY = 'REFRESH_DAILY';
+export const BACKGROUND_REFRESH_WEEKLY = 'REFRESH_WEEKLY';
+export const BACKGROUND_REFRESH_MANUALLY = 'REFRESH_MANUALLY';
