+++ src/features/backgrounds/settings/SourceChoiceField.svelte
