+++ src/features/backgrounds/settings/index.js
@@ -1,0 +1,5 @@
+import BackgroundSettings from './BackgroundSettings.svelte';
+
+export const id = 'BACKGROUND';
+export const label = 'Background';
+export const component = BackgroundSettings;
