+++ src/features/settings/config.js
@@ -1,6 +1,7 @@
 import * as ThemeSettings from '@features/themes/settings';
+import * as BackgroundSettings from '@features/backgrounds/settings';
 import * as QuickLinksSettings from '@features/quick-links/settings';
 import * as TagsSettings from '@features/tags/settings';
 import * as ShortcutsSettings from '@features/shortcuts/settings';
 
-export const settingsTabs = [ThemeSettings, QuickLinksSettings, TagsSettings, ShortcutsSettings];
+export const settingsTabs = [ThemeSettings, BackgroundSettings, QuickLinksSettings, TagsSettings, ShortcutsSettings];
