+++ src/lib/constants.js
@@ -6,10 +6,3 @@ export const STORAGE_KEY_TAGS = 'simple-todo-tags';
 export const TODOS_TODAY = 'TODAY';
 export const TODOS_THIS_WEEK = 'THIS_WEEK';
 export const TODOS_EVENTUALLY = 'EVENTUALLY';
-
-export const BACKGROUND_AUTOMATIC = 'AUTOMATIC';
-export const BACKGROUND_CUSTOM = 'CUSTOM';
-
-export const BACKGROUND_REFRESH_DAILY = 'REFRESH_DAILY';
-export const BACKGROUND_REFRESH_WEEKLY = 'REFRESH_WEEKLY';
-export const BACKGROUND_REFRESH_MANUALLY = 'REFRESH_MANUALLY';
